Guard against RealmConfig being null.
Partial Fix #319.
Really can't see how RealmConfig is null in these crashes so there is still an underlying issue.
This change catches the exception and logs it. There is loss of functionality if realm config is null in these
situations but not enough to crash the app - needs more investigation.