Ensure signout works
Block the deletion of local data until the web calls are done.
If the web calls fail we will still delete local data, the web will be in an invalid state but so be it.

Ensuring that when deregistering from ETH GCMs that we delete all addresses, just just the local one.