Init managers in constructor
This has no side effects but does cause the constructor to do some work.
This is necessary because when a dead app is awoken from the background none of the init calls
are executed. By initialising in the constructor we will be ready.

This initialisation runs in a worker thread, only works if a wallet already exists, and the initilisation is smart enough to only
do any work once; so it shouldn't have any side effects.