Fix message receiving bug
When the app goes to the background, onTrimMemory is called which shuts down the message receiver and setting it to null. When the app is resumed the message receiver wouldn't be resumed. That's why we couldn't receive messages when putting the app in the background. You had to kill the app and start it again to fix it