Correctly clear previous ChatActivities when opening multiple
There was an issue whereby when opening many ChatActivitities they quickly stacked up on top of each other.
singleTask was used as a fux but that did not work. The second activity would revive the old activity including its cached user.
This fix correctly clears the previous tasks whilst generating a new task.