Remove BTC from list of currencies.
This is a temporary measure because we currently limit local currencies to 2DP in the UI. This leaves BTC (and potentially other currencies)
unable to render correctly - they render as 0.00 for a wide range of ETH values. This is a design issue, but seeing as BTC may be a popular
currency, it has been disabled until we have a design fix.