Ensure new message listener isn't incorrectly detached.
If moving from one ChatActivity to another (for example when in chat, and clicking a push notification)
the second ChatActivity would be broken because its new message listener would be detached when
the first activity was destroyed.

By explicitly passing the thread ID that the activity should stop listening to changes for we avoid
this issue.