Payment confirmation (#833)
* Create a button with a spinner

Going to be used in the paymant confirmation view

* Add payment confirmation layout

* Payment confirmation generation (#835)

* Rename GasPrice to EthAndFiat

Make a wrapper round local and eth value. Can be reused

* Add totalAmount, paymentAmount and callbackId to PaymentTask

Before we had the callbackId in the bundle and we send to used to parameters in the callback. Now we only use PaymentTask. Also added the totalAmount and paymentAmount to have an easier way to render the payment info

* Make the mapToString funtion to return a String instead of Single<String>

This function doesn't need to be wrapped in a Single

Also make getLocalCurrencyExchangeRate public because we want to use it in PaymentTaskBuilder

* Add payment info to PaymentTask

* Add payment to PaymentTask when w3Transaction

* Payment confirmation view (#836)

* Add a function that returns Single<Balance>

Return Balance instead of a String. Need the info in Balance to compare

* Add PaymentConfirmationViewModel

* Add PaymentConfirmationFragment

* Payment confirmation dapp external (#837)

* Remove unnecessary network request

User already exists in PaymentTask if the address belongs to a toshi user

* Add an ImageView to the dapp layout

* Render dapp info if it's a dapp

* Pass dapp info to PaymentConfirmationFragment

* Render external payment info

* Replace payment confirmation (#838)

* Delete old payment confirmation

* Replace old payment confirmation

* Dismiss dialog in onPause

* Payment confirmation fixes (#840)

* Add elevation to the payment confirmation toolbar

* Add top margin to loading text

* Add ... at the end of the loading text

* Add style to the recipient name

* Make avatar smaller

* Set the height of the PaymentConfirmationFragment (#841)

Set the height to be MATCH_PARENT if the payment != w3 payment