Nav bar was pushed down (#940)
When switching from Dapp fragment to recent/wallet/fav/me, the nav bar would be pushed down and then go up again. This is because of the window inset. The Dapp fragment uses fitSystemWindows and the rest doesn't. Had to wrap every fragment layout with a CoordinatorLayout. This isn't optimal. Should be possible to apply insets on regular view, but I couldn't find a solution for it