Create chat (#988)
* Update new chat layout

* Remove search from UserParticipantsFragment

* Add ChatSearchActivity

* Add adapter for tabs/viewpager

* Add extension function for getting view model

* Java -> kotlin

* Search and render users

Endpoint is going to change. Using this one to have something to go afteEndpoint is going to change. Using this one to have something to go after

* Add UserV2

Create a new user, so the old user can function with all endpoints

* Add endpoints for fetching users/bots/groups

* Search for users/bots/groups

* Add observer for each search result type

* Handle orientation change

LiveData returns the result before the ViewPager has added the views

* Add click listener

* Remove spaces

* Various fixes

* Start ChatSearchActivity

* Add different view types

Add the option to view user profile or go to chat